requirements.txt
asgiref==3.11.1
Django==6.0.2
gunicorn==25.3.0
packaging==26.1
pillow==12.2.0
sqlparse==0.5.5
whitenoise==6.9.0
django-phonenumber-field==8.4.0
phonenumbers==9.0.28
